Flyway 추가 및 초기화 마이그레이션 추가

+++ api/src/main/java/com/mohajistudio/developers/api/domain/authentication/service/EmailService.java
@@ -107,7 +107,7 @@ public class EmailService {
         }
 
         LocalDateTime now = LocalDateTime.now();
-        emailVerification.setAttempts(emailVerification.getAttempts() + 1);
+        emailVerification.setAttempts((short) (emailVerification.getAttempts() + 1));
 
         emailVerification.setVerifiedAt(now);
